% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geo_columns.R
\name{add_geo_columns}
\alias{add_geo_columns}
\title{Adds geography columns from the Scottish Statistics Postcode Lookup}
\usage{
add_geo_columns(
  input_data_formatted,
  sspl_lookup,
  datazone_lookup,
  simd_lookup = NULL
)
}
\arguments{
\item{input_data_formatted}{A data frame containing the input data. Must include:
\itemize{
\item \code{postcode_formatted}: A cleaned UK postcode column.
\item \code{valid_uk_postcode}: A logical column created by the \code{clean_geo_postcode} function.
}}

\item{sspl_lookup}{A data frame containing the Scottish Statistics Postcode Lookup.}

\item{datazone_lookup}{A data frame for the data zone lookup.}

\item{simd_lookup}{(Optional) A data frame for SIMD 2020 lookup.}
}
\value{
A data frame with additional columns:
\itemize{
\item Local authority name/code
\item Health board name/code
\item Urban/rural classification
\item (Optional) SIMD 2020 quintile and decile.
}
}
\description{
This function takes a data frame along with the \code{postcode_formatted}
column and adds relevant geography code fields from the Scottish Statistics
Postcode Lookup for all Social Security Scotland official statistics
publications.
}
\details{
It is mandatory that the \code{clean_geo_postcode} function is run before the \code{add_geo_columns} function.

This function enriches an input dataset with relevant geographic information
by joining Scottish Statistics Postcode Lookup (SSPL) data. As the SSPL file
does not contain the local authority code-name and other geography
code-names, we get those code-names from the data zone 2022 lookup. From \code{datazone_lookup} the following fields are added:
\itemize{
\item \code{la_name, la_code}: Local authority area
\item \code{hb_code, hb_name}: Health board
\item \code{ur8_code, ur8_name}: Urban/rural classification 8-fold.
}

Optionally, we also get the Scottish Index of Multiple Deprivation (SIMD)
2020 columns. If \code{simd_lookup} is provided, the quintile and decile fields are added from the SIMD 2020 lookup. Note that the 2020 version of SIMD is based on the 2011 data zone code.

This function also assigns descriptive labels for unmatched postcodes:
\itemize{
\item 'Unknown - Scottish postcode' if its postcode area is identified as
Scottish postcode
\item 'Unknown - Non-Scottish postcode' if it meets the standard UK postcode format, but its postcode area is not identified as Scottish postcode
\item 'Unknown - Others' if it does not meet the standard UK postcode format.
}

The three lookup files can be obtained by using functions from sssstats package
\code{get_sspl_lookup()}; \code{get_datazone_lookup()}; \code{get_simd_lookup()}.

This function is called inside the \code{add_geography} function.
}
\examples{
\dontrun{
mock_input_data <- data.frame(postcode =
  c("AB39 2HP",
    "FK2     9BB",
    "G20 7XN",
    "G27XR",
    "CA17 9UB",
    "WA15 6NL",
    "",
    NA_character_
    )) |>
    clean_geo_postcode(postcode)

datazone_2022_lookup <- get_datazone_lookup("2022")
simd_lookup <- get_simd_lookup()
sspl_lookup <- get_sspl_lookup("filepath/to/SSPL CSV file")
add_geo_columns(mock_input_data, sspl_lookup, datazone_2022_lookup, simd_lookup, postcode_formatted)
}

}
\seealso{
\itemize{
\item \code{\link[=clean_geo_postcode]{clean_geo_postcode()}} cleans the column containing postcodes within a data frame.
\item \code{\link[=get_datazone_lookup]{get_datazone_lookup()}} gets the data zone lookup.
\item \code{\link[=get_simd_lookup]{get_simd_lookup()}} gets the Scottish Index of Multiple Deprivation lookup.
\item \code{\link[=get_sspl_lookup]{get_sspl_lookup()}} gets the Scottish Statistics Postcode Lookup.
}
}
