% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geography.R
\name{add_geography}
\alias{add_geography}
\title{Adds geography fields from the Scottish Statistics Postcode Lookup}
\usage{
add_geography(
  input_data,
  sspl_lookup,
  datazone_lookup,
  simd_lookup = NULL,
  postcode_column
)
}
\arguments{
\item{input_data}{dataset containing the postcode column}

\item{sspl_lookup}{the sspl_lookup dataframe}

\item{datazone_lookup}{the datazone lookup dataframe}

\item{simd_lookup}{the simd lookup dataframe}

\item{postcode_column}{column name containing postcode}
}
\value{
a data frame with geography fields added.
}
\description{
This function takes a data frame, and run both \code{format_geo_postcode} and
\code{add_geo_columns} functions to add relevant geography code fields from
the Scottish Statistics Postcode Lookup for all Social Security
Scotland official statistics publications.
}
\details{
Prior to \code{sssstats} version 0.2.1 which both \code{format_geo_postcode()} and
\code{add_geo_columns()} functions were created, this function did what the
combination of the latter functions do.
}
\examples{
\dontrun{
mock_input_data <- data.frame(postcode =
  c("AB39 2HP",
    "FK2     9BB",
    "G20 7XN",
    "G27XR",
    "CA17 9UB",
    "WA15 6NL",
    "",
    NA_character_
    ))

datazone_2022_lookup <- get_datazone_lookup("2022")
simd_lookup <- get_simd_lookup()
sspl_lookup <- get_sspl_lookup("filepath/to/SSPL CSV file")
add_geography(mock_input_data, sspl_lookup, datazone_2022_lookup, simd_lookup, postcode)
}

}
\seealso{
\itemize{
\item \code{\link[=format_geo_postcode]{format_geo_postcode()}} formats the column containing postcodes within a data frame.
\item \code{\link[=add_geo_columns]{add_geo_columns()}} add the geography fields from the Scottish Statistics Postcode Lookup.
\item \code{\link[=get_datazone_lookup]{get_datazone_lookup()}} gets the data zone lookup.
\item \code{\link[=get_simd_lookup]{get_simd_lookup()}} gets the Scottish Index of Multiple Deprivation lookup.
\item \code{\link[=get_sspl_lookup]{get_sspl_lookup()}} gets the Scottish Statistics Postcode Lookup.
}
}
