% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_processing_time.R
\name{add_processing_time}
\alias{add_processing_time}
\title{add_processing_time Adds a column to df containing number of working days between two dates.}
\usage{
add_processing_time(
  df,
  date_col_from,
  date_col_to,
  output_col_name,
  adjust = NULL
)
}
\arguments{
\item{df}{dataframe containing data to which the new column will be added}

\item{date_col_from}{column in date format to count from (unquoted)}

\item{date_col_to}{column in date format to count to (unquoted)}

\item{output_col_name}{column to put the processing time into (unquoted)}

\item{adjust}{number of days to add to total default is NULL which will not add any days.}
}
\value{
dataframe identical to \code{df} but with an additional column (named using
\code{output_col_name}) containing the calculated processing times.
}
\description{
Adds a column processing_time which counts the number of working days
between col_date_from and date_col_to.
The function:
\itemize{
\item checks that \code{df} is a data frame;
\item verifies that the specified columns exist;
\item requires that the date columns are of type \code{Date} or \code{POSIXct};
\item constructs SSS calendar using \code{sssstats::create_sss_calendar};
\item computes business days using \code{bizdays::bizdays()}.
\item if adjust is not zero adds a number of days.
}
Errors are raised if required packages are missing, if date columns are
invalid, or if the calendar making or business‑day calculation fails.
}
\examples{
# Example data
df_example <- data.frame(
  received_date = as.Date(c("2024-01-02", "2024-01-10", "2024-02-01")),
  decision_date = as.Date(c("2024-01-05", "2024-01-15", "2024-02-10"))
)

# Add processing time without adjustment
add_processing_time(
  df = df_example,
  date_col_from = received_date,
  date_col_to = decision_date,
  output_col_name = processing_time
)

# Add processing time with a 1 day adjustment
add_processing_time(
  df = df_example,
  date_col_from = received_date,
  date_col_to = decision_date,
  output_col_name = processing_time,
  adjust = 1
)

}
