% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{adorn_financial_years}
\alias{adorn_financial_years}
\title{Adorns Sum by Financial Year to a Table}
\usage{
adorn_financial_years(data, month, financial_years)
}
\arguments{
\item{data}{A data frame}

\item{month}{Column containing months}

\item{financial_years}{Column containing financial years}
}
\description{
This function behaves like \code{\link[janitor:adorn_totals]{janitor::adorn_totals()}}, but adds totals by
financial year.
}
