% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tools.R
\name{age_on_date}
\alias{age_on_date}
\title{Calculate age in years using two dates}
\usage{
age_on_date(date_column, reference_date)
}
\arguments{
\item{date_column}{Column for Date of Birth}

\item{reference_date}{Date or column of dates to calculate age on}
}
\description{
This function calculates the age someone would be between a date of birth and a given date,
either single date or column of dates
}
