% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{bucket_other}
\alias{bucket_other}
\title{Buckets All Non-Allowed Values to "Other"}
\usage{
bucket_other(data, column, allowed_values, other = "Other")
}
\arguments{
\item{data}{A data frame}

\item{column}{Columns to be converted}

\item{allowed_values}{Vector of allowed values. These will not be converted.}

\item{other}{Value to be substituted for non-allowed values. Defaults to "Other"}
}
\description{
This function mutates a chosen column to substitute a generic value
if the current value is not in a given vector. Defaults to substituting
"Other". Useful for grouping infrequent or non-permitted values into a single
bucket.
}
