% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tools.R
\name{convert_col_date}
\alias{convert_col_date}
\title{Convert All Date Strings in a Vector from One Format to Another}
\usage{
convert_col_date(data, input_format = "\%Y-\%m", output_format = "\%B \%Y")
}
\arguments{
\item{data}{Vector of dates to be converted to new format}

\item{input_format}{Input format of date. Defaults to "\%Y-\%m"}

\item{output_format}{Output format of date. Defaults to "\%B \%Y"}
}
\description{
This function applies \code{\link[=convert_date]{convert_date()}} over a vector, defaulting to the standard Social
Scotland date format, i.e. "\%B \%Y"
}
