% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tools.R
\name{convert_date}
\alias{convert_date}
\title{Convert Date String from One Format to Another}
\usage{
convert_date(date, input_format = "\%Y-\%m", output_format = "\%B \%Y")
}
\arguments{
\item{date}{Date to be converted to new format}

\item{input_format}{Input format of date. Defaults to "\%Y-\%m"}

\item{output_format}{Output format of date. Defaults to "\%B \%Y"}
}
\description{
This function takes a string as input, parses it as a date if possible
and outputs a date in a new format, defaulting to the standard Social
Scotland date format, i.e. "\%B \%Y"
}
\examples{
convert_date("24-11")
convert_date("13/11/24", \%d/\%m/\%Y)
convert_date("13 November 2024", \%d \%B \%Y, \%Y-\%m)
}
