% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{convert_to_age_band}
\alias{convert_to_age_band}
\title{Adds a Coloumn of Age Bands to a Data Frame}
\usage{
convert_to_age_band(data, age_col, age_breaks, na = "Unknown")
}
\arguments{
\item{data}{A data frame}

\item{age_col}{Column containing age data}

\item{age_breaks}{A vector of breaks in age bands}

\item{na}{Value to convert nas to. Defaults to "Unknown".}
}
\description{
This function takes data frame and adds a column of age bands based on a column
of ages, according to a given age band specification.
}
