% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{create_age_band_spec}
\alias{create_age_band_spec}
\title{Converts List of Ages to Age Band Specification}
\usage{
create_age_band_spec(ages)
}
\arguments{
\item{ages}{Vector of breaks in age bands}
}
\description{
Converts List of Ages to Age Band Specification
}
