% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_tables.R
\name{diff_column}
\alias{diff_column}
\title{Take the Difference of a Matching Column in Two Tables}
\usage{
diff_column(table_1, table_2, column)
}
\arguments{
\item{table_1}{First table}

\item{table_2}{Second table}

\item{column}{column to consider}
}
\description{
This function returns a vector of differences corresponding to a specified column
in two data frames. It will only consider rows present in both data
sets based on the first column.
}
