% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_tables.R
\name{diff_tables}
\alias{diff_tables}
\title{Take the Difference of All Matching Columns in Two Tables}
\usage{
diff_tables(table_1, table_2)
}
\arguments{
\item{table_1}{First table}

\item{table_2}{Second table}
}
\description{
This function returns a table of differences for any matching numeric columns
in two dataframes. It will only consider rows and columns present in both data
sets (row selection is based on the first column).
}
