% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tables.R
\name{extract_table}
\alias{extract_table}
\title{Extract Publication Table from Raw Import}
\usage{
extract_table(data)
}
\arguments{
\item{data}{A data frame read from a publication file}
}
\description{
This function takes raw input read from a publication file and extracts the
underlying data
}
