% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tools.R
\name{financial_year}
\alias{financial_year}
\title{Convert Date into Financial Year}
\usage{
financial_year(date)
}
\arguments{
\item{date}{Date to be converted to new format}
}
\description{
This function takes a date object as input, and outputs string corresponding to
financial year starting at 1 April.
}
