% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode.R
\name{format_postcode}
\alias{format_postcode}
\title{Format the column containing postcodes}
\usage{
format_postcode(input_data, postcode_column)
}
\arguments{
\item{input_data}{A data frame containing the postcode column.}

\item{postcode_column}{The unquoted name of the column in \code{input_data} that contains postcodes.}
}
\value{
A data frame with two additional columns: \code{valid_uk_postcode} and \code{formatted_postcode}
}
\description{
This function takes a data frame and a column containing postcodes, formats
the postcodes, validates them against a regex pattern for the standard UK
postcode format, and returns the original data with two additional columns:
\itemize{
\item \code{postcode_formatted}: The cleaned and properly spaced postcode (or \code{NA} if invalid).
\item \code{valid_uk_postcode}: A logical flag indicating whether the formatted postcode meets the standard UK postcode format or not.
}
}
\details{
The UK postcodes are liable to change over time and thus this function does
not confirm whether a given postcode does exist or not. Instead, this
function only assess whether the given postcode is consistent with the regex
pattern for the standard UK postcode format. This is not intended as a
comprehensive stand-alone postcode cleaning function, but provides the
minimum required for joining with the Scottish Statistics Postcode Lookup.

This function:
\enumerate{
\item Converts postcodes to uppercase and trims whitespace.
\item Removes extra spaces and inserts a space before the last three characters if length >= 5.
\item Validates the formatted postcode against a UK postcode regex pattern.
}
This function is called inside the \code{add_geography()} function.

For further explanation on the structure of a typical UK postcode:
https://www.ons.gov.uk/methodology/geography/ukgeographies/postalgeography.

For further explanation on the structure of a typical Scottish postcode:
https://www.nrscotland.gov.uk/publications/geography-postcode-information-note/.
}
\examples{
 \dontrun{library(dplyr)
test_data <- tibble(
  id = 1:4,
  postcode = c("AB1 0AA", "ab10ab", "INVALID", NA)
)
format_postcode(test_data, postcode)
}
}
\seealso{
\itemize{
\item \code{\link[=add_geo_columns]{add_geo_columns()}} add the geography fields from the Scottish Statistics Postcode Lookup.
}
}
