% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_add_columns.R
\name{geo_add_columns}
\alias{geo_add_columns}
\title{geo_add_columns adds Geographic Columns to a Dataset}
\usage{
geo_add_columns(
  input_data_formatted,
  sspl_lookup,
  datazone_lookup,
  simd_lookup = NULL
)
}
\arguments{
\item{input_data_formatted}{A data frame containing the input data. Must include:
\itemize{
\item \code{postcode_formatted}: A cleaned UK postcode column.
\item \code{valid_uk_postcode}: A logical column created by geo_clean_postcode.
}}

\item{sspl_lookup}{A data frame containing the Scottish Statistics Postcode Lookup.}

\item{datazone_lookup}{A data frame for Data Zone 2022 lookup gives the la_name.}

\item{simd_lookup}{(Optional) A data frame for SIMD 2020 lookup.}
}
\value{
A data frame with additional columns:
\itemize{
\item Local authority name/code
\item Health board name/code
\item Urban/rural classification
\item SIMD quintile and decile (if available)
}
}
\description{
geo_add_columns adds Geographic Columns to a Dataset
}
\details{
The geo_ prefix indicates that the function is designed to be used after
cleaning the postcode with geo_clean_postcode.
This function contains code that used to be in add_geography.
It has been removed from add_geography to allow cleaning to be carried out
separately to the addition of geography columns such as local authority name.

This function enriches an input dataset with geographic information by joining
Scottish Statistics Postcode Lookup (SSPL) data,
Data Zones and optional SIMD. It also assigns descriptive labels for
unmatched postcodes.
\itemize{
\item From \code{datazone_lookup} the following fields are added:
\itemize{
\item \code{la_name ,la_code}: Local authority area
\item \code{hb_code, hb_name}: Health Board
\item \code{ur8_code, ur8_name}: Urban rural classification 8-fold
}
\item If \code{simd_lookup} is provided, SIMD quintile and decile fields are added.
\item Scottish postcode areas are identified dynamically (excluding "CA").
\item For unmatched postcodes, descriptive labels are assigned based on whether the
postcode is Scottish, non-Scottish, or invalid.
}

The three lookup files can be obtained by using functions from sssstats package
get_sspl_lookup();get_datazone_lookup();get_simd_lookup()
}
\examples{
\dontrun{
my_data_with_geography <- geo_add_columns(
  input_data_formatted = my_data,
  sspl_lookup = sspl_lookup,
  datazone_lookup = dz_lookup,
  simd_lookup = simd_lookup
)
}

}
