% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_clean_postcode.R
\name{geo_clean_postcode}
\alias{geo_clean_postcode}
\title{geo_clean_postcodes Cleans and Validates UK Postcodes}
\usage{
geo_clean_postcode(input_data, postcode_column)
}
\arguments{
\item{input_data}{A data frame containing the postcode column.}

\item{postcode_column}{The unquoted name of the column in \code{input_data} that contains postcodes.}
}
\value{
A data frame with all original columns and valid_uk_postcode, formatted_postcode
}
\description{
geo_clean_postcodes Cleans and Validates UK Postcodes
}
\details{
The geo_ prefix indicates that the function is designed to be used before
adding geography columns with geo_add_columns.
This function contains postcode cleaning that used to be in add_geography.
It has been removed from add_geography to allow cleaning to be carried out
separately to the addition of geography columns such as local authority name.
It is not intended as a comprehensive stand-alone postcode cleaning function
but provides the minimum required for joining with sspl.

This function takes a data frame and a column containing UK postcodes,
cleans and formats the postcodes, validates them against a UK postcode
regex, and returns the original data with two additional columns:
\itemize{
\item \code{postcode_formatted}: The cleaned and properly spaced postcode (or \code{NA} if invalid).
\item \code{valid_uk_postcode}: A logical flag indicating whether the postcode is valid.
}
The function:
\enumerate{
\item Converts postcodes to uppercase and trims whitespace.
\item Removes extra spaces and inserts a space before the last three characters if length >= 5.
\item Validates the formatted postcode against a UK postcode regex pattern.
}
This function is called inside the add_geography function.
}
\examples{
 \dontrun{library(dplyr)
test_data <- tibble(
  id = 1:4,
  postcode = c("TS1 4AX", "ts14ax", "INVALID", NA)
)
geo_clean_postcode(test_data, postcode)
}
}
