% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tables.R
\name{import_tables}
\alias{import_tables}
\title{Import and Extract Publication Tables}
\usage{
import_tables(filepath)
}
\arguments{
\item{filepath}{Excel file containing publication tables}
}
\description{
This function takes an excel file of publication tables, reads in the data and
extracts the data tables.
}
