% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_all_low_negative}
\alias{make_all_low_negative}
\title{Converts Low Values to -100 Over a Table}
\usage{
make_all_low_negative(
  table,
  convert = "\\\\[low\\\\]",
  across = tidyselect::everything()
)
}
\arguments{
\item{table}{A data frame}

\item{convert}{String indicating low value, to be converted}

\item{across}{Columns to convert across. Takes tidyselect specification.
Defaults to \code{everything()}}
}
\description{
This function applies \code{\link[=make_low_negative]{make_low_negative()}} across a whole table.
Defaults to converting the string "[low]".
}
