% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_all_number_percent}
\alias{make_all_number_percent}
\title{Converts Number to Percent Over a Table}
\usage{
make_all_number_percent(table, across = tidyselect::contains("Percentage"))
}
\arguments{
\item{table}{A data frame}

\item{across}{Columns to convert across. Takes tidyselect specification.
Defaults to \code{contains("Percentage")}}
}
\description{
This function applies \code{\link[=make_number_percent]{make_number_percent()}} across a whole table.
}
