% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_all_supp_negative}
\alias{make_all_supp_negative}
\title{Converts Suppressed Values to -1 Over a Table}
\usage{
make_all_supp_negative(
  table,
  suppress = "\\\\[c\\\\]",
  across = tidyselect::everything()
)
}
\arguments{
\item{table}{Table to have values converted}

\item{across}{Columns to convert across. Takes tidyselect specification.
Defaults to \code{everything()}}

\item{convert}{String indicating suppression, to be converted}
}
\description{
This function applies \code{\link[=make_supp_negative]{make_supp_negative()}} across a whole table.
Defaults to converting the string "\link{c}".
}
