% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_low_negative}
\alias{make_low_negative}
\title{Converts Low Value to -100}
\usage{
make_low_negative(value, convert = "\\\\[low\\\\]")
}
\arguments{
\item{value}{Value to be converted}

\item{convert}{String indicating low value, to be converted}
}
\description{
This function take a value and if it is labelled low converts it to the string
"-100".
Defaults to converting the string "[low]".
This is useful for preparing suppressed tables for excel output and
manipulating numeric columns with suppression.
}
