% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_number_percent}
\alias{make_number_percent}
\title{Converts Number to Percent}
\usage{
make_number_percent(value)
}
\arguments{
\item{value}{Value to be converted}
}
\description{
This function takes a string and, if it can be parsed as a number, divides it
by 100 to return a percentage. Otherwise it returns the string.
}
