% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{make_string_numeric}
\alias{make_string_numeric}
\title{Converts String to Numeric If Possible}
\usage{
make_string_numeric(value)
}
\arguments{
\item{value}{Value to be converted}
}
\description{
This function takes a string and, if it can be parsed as a number, returns
a numeric. Otherwise it returns the string.
}
