% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_tools.R
\name{mround}
\alias{mround}
\title{Rounds to the nearest desired multiple.}
\usage{
mround(x, multiple)
}
\arguments{
\item{x}{Number to be rounded}

\item{multiple}{Multiple to be rounded to}
}
\description{
This function behaves like Excel's \code{mround}.
}
