% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_date.R
\name{pretty_date}
\alias{pretty_date}
\title{Convert Date to Pretty Output Format}
\usage{
pretty_date(date)
}
\arguments{
\item{date}{Date to be converted to pretty format}
}
\description{
This function takes a string as input, parses it as a date if possible
and outputs a date in the standard Social Security Scotland publication format
}
\examples{
pretty_date()
}
\keyword{date}
