% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_types.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace NA Over a Table}
\usage{
replace_all_na(table, replace, across = tidyselect::everything())
}
\arguments{
\item{table}{A data frame}

\item{replace}{Value to replace NA with}

\item{across}{Columns to convert across. Takes tidyselect specification.
Defaults to \code{everything()}}
}
\description{
This function applies \code{\link[=replace_na]{replace_na()}} across a whole table.
}
