% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_tools.R
\name{round_and_suppress}
\alias{round_and_suppress}
\title{Rounds whole data frame to the nearest multiple  and suppresses any values below 5 a limit}
\usage{
round_and_suppress(data, multiple = 5, suppression_value = -1)
}
\arguments{
\item{data}{A data frame to be rounded and suppressed}

\item{multiple}{Multiple to which to round table}

\item{suppression_value}{Value to convert suppressed values to. Defaults to -1
for ease of output to Excel.}
}
\description{
This function take a data frame, rounds every value to the nearest given multiple,
then converts any values that round to zero to a given suppression value.
}
