% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{transpose_data}
\alias{transpose_data}
\title{Transposes a Data Frame}
\usage{
transpose_data(data, pivot_column = 1, names_to = "name")
}
\arguments{
\item{data}{A data frame to transpose.}

\item{pivot_column}{Column to be turned into column headings. Defaults to the
first column in the data frame.}

\item{names_to}{A character vector specifying the name for the column created
from the column names.}
}
\description{
This function take a data frame or tibble and transposes it, turning columns to
rows and rows to columns
}
