% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geography.R
\name{add_geography}
\alias{add_geography}
\title{Adds geography fields from the Scottish Statistics Postcode Lookup}
\usage{
add_geography(
  input_data,
  sspl_lookup,
  datazone_lookup,
  simd_lookup,
  postcode_column
)
}
\arguments{
\item{input_data}{dataset containing the postcode column}

\item{sspl_lookup}{the sspl_lookup dataframe}

\item{datazone_lookup}{the datazone lookup dataframe}

\item{simd_lookup}{the simd lookup dataframe}

\item{postcode_column}{column name containing postcode}
}
\value{
data frame with geography fields added
}
\description{
Adds relevant geography code fields from the Scottish Statistics Postcode
Lookup for all Social Security Scotland official statistics publications.

As this lookup does not contain the local authority name and other geography
codename required for official statistics publications, we need to join
the local authority names from the data zone 2022 lookup.

To add Scottish Index of Multiple Deprivation columns the 2011 data zone
code is also needed and these are obtained from the Scottish Statistics
Postcode Lookup file.

After cleaning the postcodes, the geography fields are added to the
target data frame by joining on the postcode column.

There are some postcodes that cannot be matched to any version of the
Scottish Statistics Postcode lookup, and thus those postcodes cannot be
assigned to the relevant geography area (e.g. a local authority area etc.)
Therefore, postcodes would be assigned to:
\itemize{
\item 'Unknown - Scottish postcode' if its postcode area is identified as
Scottish postcode
\item 'Unknown - Non-Scottish postcode' if it meets the UK standard postcode,
but its postcode area is not identified as Scottish postcode
\item 'Unknown - Others' if it does not meet the UK standard postcode.
}

For further explanation on the structure of a typical UK postcode:
https://www.ons.gov.uk/methodology/geography/ukgeographies/postalgeography.

For further explanation on the structure of a typical Scottish postcode:
https://www.nrscotland.gov.uk/publications/geography-postcode-information-note/.

The three lookup files can be obtained by using functions from sssstats
package: \code{get_sspl_lookup}; \code{get_datazone_lookup} (2022) ; \code{get_simd_lookup}.
}
\examples{
\dontrun{
mock_input_data <- data.frame(postcode =
  c("AB39 2HP",
    "FK2     9BB",
    "G20 7XN",
    "G27XR",
    "CA17 9UB",
    "WA15 6NL",
    "",
    NA_character_
    ))

datazone_2022_lookup <- get_datazone_lookup("2022")
simd_lookup <- get_simd_lookup()
sspl_lookup <- get_sspl_lookup("filepath/to/SSPL CSV file")
add_geography(mock_input_data, sspl_lookup, datazone_2022_lookup, simd_lookup, postcode)
}

}
\seealso{
\itemize{
\item \code{\link[=get_datazone_lookup]{get_datazone_lookup()}} gets the data zone lookup.
\item \code{\link[=get_simd_lookup]{get_simd_lookup()}} gets the Scottish Index of Multiple Deprivation lookup.
\item \code{\link[=get_sspl_lookup]{get_sspl_lookup()}} gets the Scottish Statistics Postcode Lookup.
}
}
