% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tools.R
\name{create_sss_calendar}
\alias{create_sss_calendar}
\title{Create standard calendar with Scottish bank holidays}
\usage{
create_sss_calendar(date_from = "2018-01-01", date_to = "2070-01-01")
}
\arguments{
\item{date_from}{Calendar start date}

\item{date_to}{Calendar end date}
}
\description{
This function generates a calendar using Scottish bank holidays for use with
\code{bizdays}.
}
