% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_lookup.R
\name{get_simd_lookup}
\alias{get_simd_lookup}
\title{Creates the Scottish Index of Multiple Deprivation 2020 (SIMD) lookup}
\usage{
get_simd_lookup()
}
\value{
A data frame.
}
\description{
Uses data.gov.scot to get the SIMD (2011 data zone based) lookup
required for the Social Security Scotland official statistics publications.

The previous data platform returned \code{ref_area} not \code{geography_code}. A
column called \code{ref_area} has been made to prevent breaking existing functions.
}
\examples{
\dontrun{
simd_lookup <- get_simd_lookup()
}
}
\seealso{
\itemize{
\item \code{\link[=get_datazone_lookup]{get_datazone_lookup()}} gets the data zone lookup.
\item \code{\link[=get_sspl_lookup]{get_sspl_lookup()}} gets the Scottish Statistics Postcode Lookup.
}
}
