% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_lookup.R
\name{get_simd_lookup}
\alias{get_simd_lookup}
\title{get_simd_lookup Creates simd lookup}
\usage{
get_simd_lookup()
}
\value{
data frame
}
\description{
downloads and cleans Scottish Index of Multiple Deprivation data from https://statistics.gov.scot
}
