% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_lookup.R
\name{get_simd_lookup}
\alias{get_simd_lookup}
\title{get_simd_lookup Creates simd lookup}
\usage{
get_simd_lookup()
}
\value{
data frame
}
\description{
Uses opendatascot package to get the simd lookup required
for the Social Security Scotland official statistics publications.
}
