% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sspl_lookup.R
\name{get_sspl_lookup}
\alias{get_sspl_lookup}
\title{get_sspl_lookup Read in the Scottish Statistics Postcode Lookup}
\usage{
get_sspl_lookup(folder_path)
}
\arguments{
\item{folder-path}{the path to the folder containing the lookup.}
}
\value{
A data frame.
}
\description{
Currently, the Scottish Statistics Postcode Lookup is available on this
webpage: https://www.nrscotland.gov.uk/publications/scottish-statistics-postcode-lookup/
-- future plans may include making this lookup available on https://statistics.gov.scot/home.

The Scottish Statistics Postcode Lookup should be used for all official
statistical production to ensure geographic consistency across all official
statistics and to support the implementation of the Government Statistical
Service Geography Policy.
}
