% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sssstats-package.R
\docType{package}
\name{sssstats-package}
\alias{sssstats}
\alias{sssstats-package}
\title{sssstats: Functions for Social Security Scotland Statisticians}
\description{
Useful helper functions designed for use in Social Security Scotland Statistics team.
}
\author{
\strong{Maintainer}: Izaak Jephson \email{izaak.jephson@socialsecurity.gov.scot}

}
\keyword{internal}
